Convert LocalRepositoryModel to GitService tests
Move LocalRepositoryModel construction tests to new home.
